Various improvements to linting and organization
Formatting
