+++ linter/Taskfile.yml
@@ -1,7 +1,7 @@
 version: '3'
 
 env:
-  NAME: NUTC-client
+  NAME: NUTC-linter
 
 tasks:
   deps:
