+++ linter/src/crow/crow.cpp
@@ -1,0 +1,38 @@
+#include "crow.hpp"
+
+namespace nutc {
+namespace crow {
+
+std::thread
+get_server_thread()
+{
+    std::thread server_thread([]() {
+        namespace crow = ::crow;
+        ::crow::SimpleApp app;
+        CROW_ROUTE(app, "/")
+        ([&](const crow::request& req) {
+            log_i(main, "Registered");
+
+            if (!req.url_params.get("uid")) {
+                log_e(main, "No uid provided");
+                return crow::response(400);
+            };
+            if (!req.url_params.get("algo_id")) {
+                log_e(main, "No algo_id provided");
+                return crow::response(400);
+            }
+            std::string uid = req.url_params.get("uid");
+            std::string algo_id = req.url_params.get("algo_id");
+
+            spawning::spawn_client(uid, algo_id);
+
+            std::this_thread::sleep_for(std::chrono::milliseconds(1000));
+            return crow::response();
+        });
+        app.port(8080).run();
+    });
+    return server_thread;
+}
+
+} // namespace crow
+} // namespace nutc
