+++ linter/src/crow/crow.hpp
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "logging.hpp"
+#include "spawning/spawning.hpp"
+
+#include <crow/app.h>
+
+#include <thread>
+
+namespace nutc {
+namespace crow {
+
+std::thread get_server_thread();
+
+}
+} // namespace nutc
