+++ linter/test/src/NUTC-linter_test.cpp
@@ -1,5 +1,7 @@
 #include <gtest/gtest.h>
 
+// #include "main.cpp"
+
 // Demonstrate some basic assertions.
 TEST(HelloTest, BasicAssertions)
 {
